back/requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.8
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-cors-headers==3.2.0
django-extensions==2.2.5
django-timezone-field==4.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.3.11
requests==2.22.0
rest-framework-generic-relations==1.2.2
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
python-http-client==3.2.4
sendgrid==6.1.1
django-sendgrid-webhook==0.3.0
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-crispy-forms==1.8.1
